+++ example/ios/Runner/AppDelegate.h
@@ -1,7 +1,6 @@
 #import <UIKit/UIKit.h>
 #import <Flutter/Flutter.h>
-#import "STPAddSourceViewController.h"
 
-@interface AppDelegate : FlutterAppDelegate<STPAddSourceViewControllerDelegate>
+@interface AppDelegate : FlutterAppDelegate
 
 @end